refactor: rename rulests + support shorthands
