+++ src/cli/commands/__tests__/lint.test.ts
@@ -15,8 +15,8 @@ const validNestedRulesetPath = resolve(__dirname, '__fixtures__/ruleset-extends-
 const invalidNestedRulesetPath = resolve(__dirname, '__fixtures__/ruleset-extends-invalid.yaml');
 const validRulesetConfigPath = resolve(__dirname, '__fixtures__/config.ruleset.yml');
 const invalidRulesetConfigPath = resolve(__dirname, '__fixtures__/config.ruleset.invalid.yml');
-const standardOas3RulesetPath = resolve(__dirname, '../../../rulesets/oas3/ruleset.json');
-const standardOas2RulesetPath = resolve(__dirname, '../../../rulesets/oas2/ruleset.json');
+const standardOas3RulesetPath = resolve(__dirname, '../../../rulesets/oas3/index.json');
+const standardOas2RulesetPath = resolve(__dirname, '../../../rulesets/oas2/index.json');
 
 /*
  * These tests currently do not assert stderr because it doesn't seem to be
