+++ src/rulesets/oas/__tests__/example-value-or-externalValue.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../ruleset.json';
+import * as ruleset from '../index.json';
 
 describe('example-value-or-externalValue', () => {
   const s = new Spectral();
