+++ src/rulesets/oas/__tests__/operation-description.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../ruleset.json';
+import * as ruleset from '../index.json';
 
 describe('operation-description', () => {
   const s = new Spectral();
