+++ src/rulesets/oas/__tests__/operation-operationId-valid-in-url.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../ruleset.json';
+import * as ruleset from '../index.json';
 
 describe('operation-operationId-valid-in-url', () => {
   const s = new Spectral();
