+++ src/rulesets/oas2/__tests__/host-not-example.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../ruleset.json';
+import * as ruleset from '../index.json';
 
 describe('host-not-example', () => {
   const s = new Spectral();
