+++ src/rulesets/oas3/__tests__/api-servers.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../ruleset.json';
+import * as ruleset from '../index.json';
 
 describe('api-servers', () => {
   const s = new Spectral();
