+++ src/rulesets/oas3/index.ts
@@ -2,4 +2,4 @@ import { readRulesFromRulesets } from '../reader';
 
 export { commonOasFunctions as oas3Functions } from '../oas';
 
-export const rules = async () => readRulesFromRulesets(require.resolve('./ruleset.json'));
+export const rules = async () => readRulesFromRulesets(require.resolve('./index.json'));
