+++ src/rulesets/path.ts
@@ -1,8 +1,14 @@
 import * as path from '@stoplight/path';
 import { PROJECT_ROOT } from '../consts';
 import { isURL } from '../fs/reader';
+import { rulesetsMap } from './map';
 
 export function resolvePath(from: string, to: string) {
+  const mapped = rulesetsMap.get(to);
+  if (mapped !== void 0) {
+    to = mapped;
+  }
+
   if (isURL(to) || path.isAbsolute(to)) {
     return to;
   }
